# blur and spread speedup + fix