… examples/blur.png
… examples/shadow.png
… tests/images/imageblur20.png
… tests/images/imageblur20oob.png
… tests/images/maskblur20.png