before
images: `blur ............................ 1113.121 ms   1122.303 ms    ±9.992     x5`
masks: `blur ............................. 386.647 ms    389.918 ms    ±2.851    x13`

after
images: `blur ............................. 834.241 ms    839.677 ms    ±8.135     x6`
masks: `blur ............................. 242.845 ms    244.885 ms    ±1.822    x21`