@guzba 2021-02-27
spread before:
`spread .......................... 1570.692 ms   1618.245 ms   ±49.552     x4`

spread after:
`spread ........................... 668.253 ms    718.283 ms   ±44.109     x8`